🚚 Split steps "la candidature lauréate notifiée" + "candidature éliminée notifiée" dans les répertoires dédiées
